requirements.txt
pyahocorasick==1.4.4
requests==2.26.0
gevent==21.12.0
jieba==0.42.1
six==1.15.0
gensim==4.1.2
matplotlib==3.5.1
Flask==2.1.2
numpy==1.19.5
tensorflow-addons==0.13.0
tensorflow-gpu==2.5.0
py2neo==2020.1.1
tqdm==4.62.3
pandas==1.3.2
termcolor==1.1.0
wechaty==0.8.39
scikit_learn==1.0.2